Move graph export functions into cloudkeeper
